+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,14 +1,14 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
 import static java.util.stream.Collectors.*;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.dto.OrderLineItemSaveRequest;
-import kitchenpos.dto.OrderResponse;
-import kitchenpos.dto.OrderSaveRequest;
-import kitchenpos.dto.OrderChangeOrderStatusRequest;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.dto.OrderLineItemSaveRequest;
+import kitchenpos.order.dto.OrderResponse;
+import kitchenpos.order.dto.OrderSaveRequest;
+import kitchenpos.order.dto.OrderChangeOrderStatusRequest;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.repository.OrderRepository;
 import org.springframework.stereotype.Service;
