+++ src/main/java/kitchenpos/order/domain/JpaOrderDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderDao extends JpaRepository<Order, Long>, OrderDao {
