+++ src/main/java/kitchenpos/order/domain/JpaOrderLineItemDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderLineItemDao extends JpaRepository<OrderLineItem, Long>, OrderLineItemDao {
