+++ src/main/java/kitchenpos/order/domain/OrderDao.java
@@ -1,6 +1,4 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Order;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
