+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemResponse {
