+++ src/main/java/kitchenpos/order/dto/OrderSaveRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
 import java.util.List;
