+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
 import static kitchenpos.common.fixture.MenuFixtures.generateMenu;
 import static kitchenpos.common.fixture.MenuGroupFixtures.generateMenuGroup;
@@ -15,20 +15,20 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.menugroup.domain.MenuGroupDao;
-import kitchenpos.dao.OrderDao;
+import kitchenpos.order.domain.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.product.domain.ProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.product.domain.Product;
-import kitchenpos.dto.OrderChangeOrderStatusRequest;
-import kitchenpos.dto.OrderLineItemSaveRequest;
-import kitchenpos.dto.OrderResponse;
-import kitchenpos.dto.OrderSaveRequest;
+import kitchenpos.order.dto.OrderChangeOrderStatusRequest;
+import kitchenpos.order.dto.OrderLineItemSaveRequest;
+import kitchenpos.order.dto.OrderResponse;
+import kitchenpos.order.dto.OrderSaveRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
